Move settings actions to their own file
